Rename tree subpackage to auto
