+++ src/qibocal/auto/execute.py
