+++ src/qibocal/auto/graph.py
