+++ src/qibocal/auto/history.py
