+++ src/qibocal/auto/operation.py
