+++ src/qibocal/auto/task.py
