+++ src/qibocal/auto/validate.py
